Added check for HttpException in Message so the bot isn't stopped when Discord is unreachable for a short amount of time
